Phase management upgrade step 3: train limits.
<Phase><Trains> now has a new attribute 'limitStep', which starts at 1.
<Company[Type]><Trains> now lists the number of trains per step, not per
phase.
These changes have been applied to 1835 ONLY. Temporary code ensures
that the old way still works for other games.
    
<Phase> has another new optional attribute 'realName', intended to have
the 'official' name of a phase, as in the physical game.
    
1835 phase example:
  <Phase name="4" realName="2">
    <Trains limitStep="2" rusted="2"/>
  </Phase>
In the game status window this phase is now displayed as "4 [2]".
The major CompanyType now has
  <Trains limit="4,3,2"/>
    
This change also fixes the 1835 bug that the train limit was reduced to
2 at buying the 4+4-train in stead of the 5-train.