Fix: Buying Coalfields right is now refused if the company does not have enough cash.
Reported by Charles Strong.
This is the easy fix. In fact, the option to buy this right should not be offered at all
if the company lacks cash, but fixing that is a bit more complex and omitted for now.
(cherry picked from commit 8dd8adf3e2da0056bb7b254f3a0d796288d73b1d)