Fix: BA home token is now laid in its first OR.
This fixes the problem that, if the BA has floated but not yet operated,
its home token is already on the map and must find a place when another 
player lays a green XX tile on the BA home hex.  
In this case, the wrong player was asked for that place.