Feature 3454269 (background music) available - Some details on the design/implementation
Rails is now capable of playing background music depending on the
context (phase, round type). So, as for SimTex's 1830, if a new phase
begins (or in case of a OR-SR switch), rails plays a different mp3 in
the background. As a result, experiencing/triggering context changes
(eg, rusting others' trains) is now much more fun.

The end user can configure which mp3 track is used for which context in
the configuration window (new tab "Sound"). There, he can also define
assignment of mp3 tracks to several contexts (e.g., the same SR music
independent of the phase - as it is done in SimTex's 1830).

For copyright reasons, no default mp3 tracks are included in rails.
During testing, I had configured rails to access SimTex's mp3 files but,
as I understand, rails would not be allowed to redistribute them.
Therefore, the end user has to provide mp3 files on his own.

Note however, that I couldn't commit the change yet due to missing
permissions. But Brett has already been contacted for that.

==========================================
Some more details about the implementation
==========================================

Playing mp3 is based on the lib JLayer which is now added to the rails
lib. This should be ok for all of you as:
- lib licencse is LGPL
- lib size is only 100kb

All the logic regarding the background music is found in
rails.sound.BackgroundMusicManager. Some comments about the most
important design decisions:
- Manager is a static class - there may only be one background music at
a time.
- Manager is notified of context changes by adding notification calls
from the context-changing parts of the game:
   - rails.game.Phase
   - rails.game.StockRound
   - rails.game.OperatingRound
  (I would have loved to subscribe to such kind of information but,
since there are no such interfaces available and you apparently aim at
refactoring for 2.0, this was not possible.)
- Manager is started/configured by calls from rails.game.Game

I hope this is ok for all of you. If you find bugs in the new code,
don't hesitate to contact me.