Fixed retrieval of default game options (if options pane was not opened)
Game options have to be read from GamesList.xml and not from the
respective game's Games.xml GameOption tags.
Before the fix, this was not the case if the options pane had not been
opened (which is the moment at which the GamesList's default options
were loaded).

This fix also solves Bug 3448429, as the reported missing 18EU route
calculation was due to inconsistent defaulting between GamesList
(default=suggest) and Games (default=deactivate). Now, the defaults
are always taken from GamesList.