Added handling of concurrent hex map paints
The key blocks / methods are now synchronized on the hex map (acting
as the monitor). This became necessary after observing very rare
paintComponent calls with invalid clip rectangles (out of JComponent
bounds), leading to out-of-bounds paints of the background map.

After this addition, no further occurrence of this symptom could be
observed.