Transformed fit-to zoom from one-time zoom to permanent auto-zoom
The fit-to-xyz zoom options are now check box menu items. The selected
option is permanent in the sense that MapPanel resizing will lead
automatically lead to the selected/active fit-to-xyz option.

In addition, added a new map config option to define what the default
zoom option is when starting a game.

Furthermore, fixed a legacy issue with initial mapImage resizing that
had been marked with FIXME.