Prohibited resizing of OR's message panel (and ORPanel to some extent)
Frequent resizing by the message panel and the OR panel caused the
map panel to being resized - which was not desirable in case of
activated auto-zoom.

Message Panel is now fixed to the height of two lines. The additional
vertical scrollbar allows for text containing more than two lines.

ORPanel used to resize itself for each set-revenue step as the size
of the revenue input box did not match the one of the revenue
read-only field. These sizes are now aligned.
(ORPanel still resizes if companies are added/removed from the list.)