Excluded layable tiles from upgrade panel if no valid rotation exists
Before, selecting a tile of the upgrade panel could lead to the message
"This tile cannot be laid in a valid orientation."

Now, tiles are excluded from the upgrade panel selection if such message
would be issued upon clicking on them. Stated differently, every tile
listed in the upgrade panel can now be laid by the user.