Created new framework for playing music and sound effects
Playing sounds (both bgm / sfx) is disabled by default.

Background music (BGM) is configurable on a per-round-type and
per-phase basis. (currently 5 round types supported)

There are 15 sound effects (SFX) types, of which
- one is parameterizable (train type for buy train)
- one is played for a varying amount of time (set revenue depending
  on the revenue - similar to RRT1)

The framework classes are found in rails.sound.

As a facade to the rest of rails, the Sound Manager is notified of
- processed events (by GameUIManager)
   - this covers close to 90% of the triggers of playing sounds
- model updates (by the respective models)

The input received by the framework is evaluated by the event
interpreter who
- triggers sound playing directly (straightforward cases)
- triggers sound context adjustment (complex case - eg. phase bgm)
    - which triggers in turn playing sounds