Player captions made variable to reflect player reordering.
Capability is game-dependent. Setting it is currently hardcoded in GameManager[_1880].setGuiParameters().
The list of players in GameManager is now an ArrayListState.
As ArrayListState does not inherit from ModelObject, a separate
StringState[] array had to be created to hold the player names for use in the GUI.
A new Cell superclass for Caption and Field allows Field to get a caption background and also centralises highlighting.