Enhanced sound framework for rails 2.0 compliance
The old BackgroundMusicManager is deleted together with all calls
to him that had been placed inside the game engine.

The new sound framework is notified as an observer to the relevant
game models. Hence, it complies to the publish/subscriber paradigm
that will be enforced in rails 2.0.

As a sidenote, GameManagerI and PublicCompanyI had to be extended
with a getter-method for required game models (getCurrentRoundModel
and getFloatedModel respectively).