Enhanced sound framework for rails 2.0 compliance
The old BackgroundMusicManager is deleted together with all calls
to him that had been placed inside the game engine.

The new sound framework is notified as an observer to the relevant
game models. Hence, it complies to the publish/subscriber paradigm
that will be enforced in rails 2.0.

As a sidenote, GameManagerI and PublicCompanyI had to be extended
with a getter-method for required game models (getCurrentRoundModel
and getFloatedModel respectively).
(cherry picked from commit 775ef65fbe5f73e84e2ecc420d7969a0597f59d3)

Conflicts:

	rails/game/GameManagerI.java
	rails/game/PublicCompanyI.java
	rails/game/specific/_1880/OperatingRound_1880.java
	src/rails/game/GameManager.java
	src/rails/game/Phase.java
	src/rails/game/RailsRoot.java
	src/rails/sound/BackgroundMusicManager.java
	src/rails/sound/SoundEventInterpreter.java
	src/rails/ui/swing/ORUIManager.java