Refactored SoundPlayer: immediate/deferred sfx play defined declaratively
Some sfx are to be played without waiting for the completion of prior
sfx (applies to selection / rotate tile sfx).

Before, the caller of the SoundPlayer had to specify whether to wait
or not.

Now, sfx to be immediately played are tagged as such in SoundConfig
so that SoundPlayer takes care of scheduling without necessiting
that information by the caller.