Rotate upgrade panel tiles to the first valid orientation
The existing logic of determining validity of orientations was re-used
(required some refactoring). This means that, once this logic is further
improved, the upgrade panel's tile orientations will also get more
accurate.

The other source of complexity is that svg tile images are not
centered and that image rotation by AOP result in additional borders
on the bottom/right. Code that is able to handle that complexity
is put to GUITile, as the former tile paint code is also there.