Rendered button panel as dockable pane (detachable from OR panel)
Change only applies if config option for dockable or window panels
is set.

Button panel's sizing is also adjusted in order to allow for both
vertical / horizontal / grid layouts of the buttons. As a
pre-condition, the buttons are all given the same size.

Added localized texts for the dock titles.