Allowed for dynamic line breaks and vertical centering of message panel
Main motivation for rework is that the dockable message text might
be put in a vertical alignment (of narrow width).

For the message panel, new behavior is that
- no horizontal scroll bar is used
- if panel's width inferior to message label's native width, a line
  break is inserted into the label text
    - technically done by using a layout manager and basing the panel's
	  preferred width on the parent scrollpane's one

An additional benefit of using the layout manager is that, for the
first time since having added the scrollbar to the panel, messages
are vertically centered.