Included "Remaining Tiles" into docking layout (as a dockable)
Change only applies to the docking layout.

Now, the remaining tiles panel is part of the ORWindow. Its
activation/positioning is done in the context of the docking
framework instead of the ORPanel menu and the JFrame.

Additionally, docking framework handles the tiles / messages window
as closable dockables. A dedicated menu option is available for
toggling the visibility.