Fixed order of initing DynamicReportWindow (pack only after layout done)
Found out issue when running on other computer:
Report window gets zero size.

This commit fixes this issue which was traced back to a premature
call of super.init to the AbstractReportWindow's pack - dyn report
window only sets up layout afterwards (against EDT rules).

This is fixed by postponing the super.init call to the point in time
when the layout is ready.