Always address the company president when a home token must be relaid.
This is a provisional fix for the issue that arises when in 1835 the BA has started but not yet operated,
and another company uses the PfB to lay a green tile in the BA home hex L6.
In this case the BA president must lay the home token in the turn of that other company.

The fix is to always address the token company president instead of the acting player during a token relay.

WARNING: this fix breaks the intended client/server architecture, because the UI
is not supposed to have direct access to the president's identity, as this is a state variable.
So this is a provisional fix only.
(cherry picked from commit d34b91f49c0ddbb9ffab3273bbe5af025652bfdd)

Conflicts:

	src/rails/ui/swing/ORUIManager.java