Fixing the bug reported by Pieter Lust for 1835.
Prussian Minor Shares still in operation got wrongly 
subtracted from the wallet of the owning player leading to
negative amounts of payouts.

This subtraction should only happen if after the companies
have operated and are folded into the Prussian Major by 
the PrussianFormationRound mechanismn. Thus the
 operationround would have been interrupted and resumed. 
 
So a check if the current OR was interrupted before the
subtraction might solve the problem.