Fixing setBuyableShares for multi share Certs
Before this fix a cert with multiple shares was shown as buyable because
th price check did not account for multiple shares per cert. The buying
would have failed neverthe less. With this fix the price per shareunit 
in the price check is taking into account the multipe shares per cert if
they are present. This only works from the IPO currently. Theres a small
chance that the wrong display might appear on shares in the pool (but
very unlikely).