Fixed a Bug in regard to shareselling while dumping the presidency.
Previously it wasnt possible to dump x shares containing the presidency.
Now it is. Befoire this fix you had to sell twice shares from the
company with the president in your hand. Now the correct amount is
available on the first sell.

Bug reported by Peter De Boer, December 2019

Conflicts:
	src/main/java/net/sf/rails/game/PlayerShareUtils.java
	src/main/java/net/sf/rails/game/StockRound.java