18 Core Setup: Changing Game load and save to add a central seed for randomization.
The need from several systems for randomization support in persistance
has been implemented. Please be advised that to achieve that the save
game format has changed. The SaveFileVersion has been raised to 12. The
old VersionID was 9. Since its easy to cover the additional information
with a single method adaption, savegames from the former version are
also accepted and transcribed.