SOH StockRound
Also StartRound improvements, and the new tile laying rule in the OR.

To make the random generator seed persistent through save/load, a new "hidden" type of GameOption was added to save the seed. Hidden means that it does not appear as an Option in the game selection window. The random generator is created in GameManager if this option exists in GameOptions.xml, and can be used via a getter.

Starting a company by buying more than 2 shares in one turn was implemented by presetting StartCompany.numberBought. The UI will no longer update this value if it is already nonzero.

Reused an 1825 run modifier to prevent trains terminating in a town.

Added moveRight and moveLeft methods to StockMarket for linear markets.