Fixed stop number issues, caused by confusing stop number and related station number.
This caused problems when stations are added by a tile upgrade, such as the Berlin yellow-to-green upgrade in 1835. The Berlin token laying issue #388 is now gone, and the original 1835/Tiles.xml could be restored.

The deprecated Stop 'legacyNumber' is rehabilitated and now just called 'number'. It is no longer equal to the related station number (which caused the confusion) but is now the integer value of the stop id. It is often used in logging.

The submitted saved game has been added to the test set with the token laid successfully, and under a different name.