1837: Implemented anti-reuse rules of certs and trains in the same round.
1. Certificates:
Major company certificates, that have been obtained by exchanging a minor or coal certificate, do not receive income if that minor or coal company has operated before *in the same operating round*.

2. Trains:
Major company trains, that have been obtained by merging a minor or coal company, do not run if that minor or coal company has operated before *in the same operating round*.

Both conditions can only occur at the start of phase 4+1 (for the KK) and phase 5 (for the coal company successors and the Ug).

It has been assumed that the mere fact of a minor having operated is sufficient to have the relevant certs and trains be registered this way, regardless whether or not any nonzero income has been gained by that minor.

These rule details are enforced by keeping a registry (HashSetState) of certificates and trains that fulfill these conditions. The registries are kept in GameManager, so that all types of Stock and Operating round instances can easily access these, and are cleared at the start of each OR.