1837: Show fixed minor prices when buyable in a stock round
Actually, the basic feature is generic, but such fixed prices (as is the case for KK and Ug minors) must (currently) be set and reset in game-specific code.

We could make it fully generic by adding some XML parameter, or by declaring the "price" attribute in a Company tag to have this effect everywhere. Not sure if this would break anything.