Allowed laying a tile on a reserved hex in Correction Mode.
This is a generic feature.

The validation methods in TileHexUpgrade can now take an additional EnumSet<Invalids> argument that specifies the checks to be bypassed. This currently only works for the HEX_RESERVED and HEX_BLOCKED enum values, but can easily be extended in TileHexUpgrade.validate().

This new feature is now used by ORUIManager, but currently only specifies the HEX_RESERVED check to be skipped.