Allowed laying a tile on a reserved hex of a company having a different president.
See previous commits.

Added a dialog to ask if the current player has got the required permission immediately after clicking the reserved hex.

This also means, that the same-presidency check before displaying the upgradeable hexes, that was introduced by the previous commit, had to be postponed; the HEX_RESERVED allowance is now always granted initially.