Final commit for 1837 beta version
Game-dependent bankruptcy handling methods collected in a new class Bankruptcy. Reason: to avoid duplicate code.

Currently, it only includes the different ways to handle the train-buying company when a player goes bankrupt but the game continues. Three styles for now: 1835 (also used by 1837), 18EU and 18Scan.

Another new class Mergers is currently only used by 1837. Both classes only have static methods.

During testing, bugs were found in 18EU and 1835, which have been fixed.

A major problem arose in 1837: it crashed when a player going bankrupt still owned minors or coal companies. This has been fixed by exchanging these minors for major certificates.
An exception to this rule applies to the Ug two-share minors U1 and U3. The exchange is only done if the bankrupt has both shares. If he has the president share only, the shares are swapped and the second minor share goes to the Pool, but will not be buyable. If the bankrupt has only the second minor share, that share will likewise go to the Pool.