18EU: added two options to add a 16th minor in case of four players
The options are: #0 on Hamburg or #16 on Frankfurt
(There is no check on the actual number of players)

A new underlying Hamburg tile -1903 with a one-slot city is used if this option is chosen. A new value ALWAYS of Access.RunThrough is used to allow a Hamburg pass-through even if it is tokened.