18VA: implementation near complete
Added default revenue calculation methods to RevenueManager, and a subclass RevenueManager_18VA to provide real stop values for the initialization of each NetworkVertex, possibly depending on train type and company properties. So RevenueManager is no longer a final class. Also the predictionValue() method in dynamic modifiers is now obsolete.

Also added a simple Revenue class, that carries two integers for normal and special revenue. For 18VA, special is used for the mine income that goes directly to the company treasury. The intention is to have revenue objects carry all revenues, but for now it is only applied to 18VA.